modified count*Bits to use size_t instead of int, and switch the order of function args so that the size is before the array. Modified subnetMaskNumber code.

+++ src/getBit.c
@@ -1,6 +1,6 @@
 #include "getBit.h"
 
-int getBit(unsigned char* byte, int index){
+int getBit(unsigned char* byte, size_t index){
   return (byte[index/8] & (1 << index%8)) ? 1 : 0;
 }
